Fixed EditButton on EditTrip
fixed #88
